+++ src/components/Progress/index.stories.tsx
@@ -1,19 +1,17 @@
-import React from "react";
-import { ComponentMeta, ComponentStory } from "@storybook/react";
+import { Meta, StoryObj } from "@storybook/react";
 
-import { localeDecorator } from "../../utils/mocks/LocaleMock";
 import getDocsPage from "../../utils/stories";
+import LocaleDecorator from "../../utils/stories/decorators/Locale";
 
-import Progress, { DATA_CY_DEFAULT } from ".";
+import { DATA_CY_DEFAULT, MemoizedProgress } from ".";
 
 const COMPONENT_NAME = "Progress";
-Progress.displayName = COMPONENT_NAME;
-//ProgressWithProps.displayName = COMPONENT_NAME;
+MemoizedProgress.displayName = COMPONENT_NAME;
 
-export default {
+const meta = {
   title: "Feedback/Progress",
-  component: Progress,
-  decorators: [localeDecorator],
+  component: MemoizedProgress,
+  decorators: [LocaleDecorator],
   parameters: {
     docs: {
       ...getDocsPage({
@@ -23,35 +21,85 @@ export default {
         },
         component: COMPONENT_NAME,
         e2eTestInfo: {
-          dataCyDefault: "progress",
+          dataCyDefault: DATA_CY_DEFAULT,
         },
       }),
+      description: {
+        component: "The Progress component is used to display the progress of a task.",
+      },
     },
   },
-} as ComponentMeta<typeof Progress>;
+} satisfies Meta<typeof MemoizedProgress>;
 
-const Template: ComponentStory<typeof Progress> = (args) => <Progress {...args} dataCy={DATA_CY_DEFAULT} />;
+export default meta;
+type Story = StoryObj<typeof meta>;
 
-export const Circular = Template.bind({});
-Circular.args = {};
+export const Circular: Story = {
+  args: {
+    type: "circular",
+  },
+};
 
-export const CircularColor = Template.bind({});
-CircularColor.args = { color: "error" };
+export const CircularColor: Story = {
+  args: {
+    ...Circular.args,
+    color: "error",
+  },
+};
 
-export const CircularDeterminate = Template.bind({});
-CircularDeterminate.args = { ...Circular.args, variant: "determinate", value: 50 };
+export const CircularDeterminate: Story = {
+  args: {
+    ...Circular.args,
+    variant: "determinate",
+    value: 50,
+  },
+};
 
-export const CircularWithLabel = Template.bind({});
-CircularWithLabel.args = { ...CircularDeterminate.args, withLabel: true };
+export const CircularInvalidValue: Story = {
+  args: {
+    ...Circular.args,
+    value: -999,
+  },
+};
 
-export const Linear = Template.bind({});
-Linear.args = { type: "Linear" };
+export const CircularWithLabel: Story = {
+  args: {
+    ...CircularDeterminate.args,
+    withLabel: true,
+  },
+};
 
-export const LinearColor = Template.bind({});
-LinearColor.args = { ...Linear.args, color: "error" };
+export const Linear: Story = {
+  args: {
+    type: "linear",
+  },
+};
 
-export const LinearDeterminate = Template.bind({});
-LinearDeterminate.args = { ...Linear.args, variant: "determinate", value: 50 };
+export const LinearColor: Story = {
+  args: {
+    ...Linear.args,
+    color: "error",
+  },
+};
 
-export const LinearWithLabel = Template.bind({});
-LinearWithLabel.args = { ...LinearDeterminate.args, withLabel: true };
+export const LinearDeterminate: Story = {
+  args: {
+    ...Linear.args,
+    variant: "determinate",
+    value: 50,
+  },
+};
+
+export const LinearInvalidValue: Story = {
+  args: {
+    ...Linear.args,
+    value: 999,
+  },
+};
+
+export const LinearWithLabel: Story = {
+  args: {
+    ...LinearDeterminate.args,
+    withLabel: true,
+  },
+};