+++ src/components/Select/index.tsx
@@ -1,3 +1,5 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
+/* eslint-disable @typescript-eslint/no-unnecessary-type-constraint */
 import React, { HTMLAttributes, SyntheticEvent, useCallback, useMemo } from "react";
 import {
   Autocomplete as MUIAutocomplete,
@@ -7,12 +9,14 @@ import {
   PopperProps as MUIPopperProps,
 } from "@mui/material";
 
-import { ISelect } from "../../types/Select";
+import { SelectData, SelectDataAllowed, SelectMultipleData, SelectProps, SelectSingleData } from "../../types/Select";
 import { getComposedDataCy, suppressEvent } from "../../utils";
 import localized, { ILocalizableProperty } from "../../utils/hocs/localized";
+import { logInfo, logWarn } from "../../utils/logger";
 
 import SelectGroup, { SELECT_GROUP_SUBPART } from "./components/Group";
 import SelectInput, { SELECT_LOADING_SUBPART } from "./components/Input";
+import SelectListBox from "./components/ListBox";
 import SelectOption, {
   SELECT_OPTION_CHECKBOX_SUBPART,
   SELECT_OPTION_LABEL_SUBPART,
@@ -38,9 +42,9 @@ export const SUBPARTS_MAP = {
   },
 };
 
-const Select = <T extends any>({
+const Select = <T extends SelectDataAllowed>({
   autoComplete = true,
-  autoSort = false,
+  autoSort = undefined,
   customOptionRendering,
   dataCy = DATA_CY_DEFAULT,
   disabled = false,
@@ -50,12 +54,13 @@ const Select = <T extends any>({
   groupBy,
   label,
   loading = false,
-  multiple = false,
+  multiple,
   onChange: externalOnChange,
   onClose: externalOnClose,
   onInputChange: externalOnInputChange,
   onScrollEnd,
   options: externalOptions,
+  optionsNumber,
   placeholder,
   popperWidth,
   required = false,
@@ -64,14 +69,17 @@ const Select = <T extends any>({
   type = "text",
   value = null,
   variant = "outlined",
-}: ISelect<T>) => {
+  virtualized = false,
+}: SelectProps<T>) => {
+  const outerWrapperDataCy = useMemo(() => getComposedDataCy(dataCy, SUBPARTS_MAP.outerWrapper), [dataCy]);
+
   const getOptionLabel = useCallback(
-    (option: T): string => (externalGetOptionLabel ? externalGetOptionLabel(option) : `${option}`),
+    (option: SelectData<T>): string => (externalGetOptionLabel ? externalGetOptionLabel(option) : `${option}`),
     [externalGetOptionLabel]
   );
 
   const isOptionSelected = useCallback(
-    (option: T, value: T): boolean => {
+    (option: SelectData<T>, value: SelectData<T>): boolean => {
       if (getOptionSelected) {
         return getOptionSelected(option, value);
       }
@@ -81,18 +89,8 @@ const Select = <T extends any>({
     [getOptionSelected]
   );
 
-  const onChange = useCallback(
-    // TODO#lb: fix any type
-    (event: any, value: any) => {
-      suppressEvent(event);
-      externalOnChange(value);
-    },
-    [externalOnChange]
-  );
-
   const onClose = useCallback(
-    // TODO#lb: fix any type
-    (event: any) => {
+    (event: SyntheticEvent) => {
       if (!externalOnClose) {
         return;
       }
@@ -104,8 +102,7 @@ const Select = <T extends any>({
   );
 
   const onInputChange = useCallback(
-    // TODO#lb: fix any type
-    (event: any, value: string) => {
+    (event: SyntheticEvent, value: string) => {
       if (!externalOnInputChange) {
         return;
       }
@@ -124,6 +121,7 @@ const Select = <T extends any>({
 
       const { clientHeight, scrollHeight, scrollTop } = listboxNode;
       const scrollEnded = clientHeight + scrollTop === scrollHeight;
+      logInfo("Select.onScroll  ", `ch ${clientHeight} / st ${scrollTop} / sh ${scrollHeight} / ${scrollEnded}`);
       if (scrollEnded) {
         onScrollEnd();
       }
@@ -131,16 +129,19 @@ const Select = <T extends any>({
     [onScrollEnd]
   );
 
-  const listboxProps = useMemo(
-    () => ({
+  const listboxProps = useMemo(() => {
+    if (virtualized) {
+      return undefined;
+    }
+
+    return {
       onScroll,
       style: {
         padding: 0,
         width: "100%",
       },
-    }),
-    [onScroll]
-  );
+    };
+  }, [onScroll, virtualized]);
 
   const options = useMemo(() => {
     let options = [...externalOptions];
@@ -149,28 +150,48 @@ const Select = <T extends any>({
     }
 
     if (groupBy) {
-      options = options.sort((one: T, another: T) => {
+      options = options.sort((one: SelectData<T>, another: SelectData<T>) => {
         const oneGroup = groupBy(one);
         const anotherGroup = groupBy(another);
         return oneGroup.localeCompare(anotherGroup);
       });
     }
 
-    if (autoSort) {
-      options = options.sort((one: T, another: T) => {
+    if (typeof autoSort === "function") {
+      options = options.sort((one: SelectData<T>, another: SelectData<T>) => {
+        return autoSort(one, another);
+      });
+    }
+
+    if (autoSort === "asc") {
+      options = options.sort((one: SelectData<T>, another: SelectData<T>) => {
         const oneLabel = getOptionLabel(one);
         const anotherLabel = getOptionLabel(another);
         return oneLabel.localeCompare(anotherLabel);
       });
     }
 
+    if (autoSort === "desc") {
+      options = options.sort((one: SelectData<T>, another: SelectData<T>) => {
+        const oneLabel = getOptionLabel(one);
+        const anotherLabel = getOptionLabel(another);
+        return anotherLabel.localeCompare(oneLabel);
+      });
+    }
+
     return options;
   }, [autoSort, externalOptions, getOptionLabel, groupBy]);
 
-  const outerWrapperDataCy = useMemo(() => getComposedDataCy(dataCy, SUBPARTS_MAP.outerWrapper), [dataCy]);
+  const listboxComponent = useMemo(() => {
+    if (!virtualized) {
+      return undefined;
+    }
+
+    return SelectListBox({ multiple, options, optionsNumber });
+  }, [multiple, options, optionsNumber, virtualized]);
 
   const isOptionSelectable = useCallback(
-    (value: T | null): boolean => !!value && options.some((option) => isOptionSelected(option, value)),
+    (value: SelectData<T> | null): boolean => !!value && options.some((option) => isOptionSelected(option, value)),
     [options, isOptionSelected]
   );
 
@@ -201,12 +222,13 @@ const Select = <T extends any>({
   );
 
   const renderCustomOption = useCallback(
-    (option: T, selected: boolean) => (customOptionRendering ? customOptionRendering(option, selected) : undefined),
+    (option: SelectData<T>, selected: boolean) =>
+      customOptionRendering ? customOptionRendering(option, selected) : undefined,
     [customOptionRendering]
   );
 
   const renderOption = useCallback(
-    (props: HTMLAttributes<HTMLLIElement>, option: T, { selected }: MUIAutocompleteRenderOptionState) => {
+    (props: HTMLAttributes<HTMLLIElement>, option: SelectData<T>, { selected }: MUIAutocompleteRenderOptionState) => {
       const { id: key } = props;
       return (
         <SelectOption
@@ -229,23 +251,60 @@ const Select = <T extends any>({
     [popperWidth]
   );
 
-  const validateValue = useCallback(
-    (value: T | T[] | null): T | T[] | null => {
-      if (multiple) {
-        const isValidMultipleValue = Array.isArray(value) && value.length > 0 && value.every(isOptionSelectable);
-        return isValidMultipleValue ? value : [];
+  if (multiple === false) {
+    const onChangeSingle = (event: SyntheticEvent, value: SelectSingleData<T>) => {
+      suppressEvent(event);
+      externalOnChange && externalOnChange(value);
+    };
+
+    let singleValue: SelectSingleData<T> = null;
+    if (!Array.isArray(value) && isOptionSelectable(value)) {
+      singleValue = value;
+    } else {
+      if (value) {
+        logWarn("Select", `Provided value is not valid '${value}'`);
       }
+    }
 
-      const isValidValue = !Array.isArray(value) && isOptionSelectable(value);
-      return isValidValue ? value : null;
-    },
-    [isOptionSelectable, multiple]
-  );
+    return (
+      <MUIAutocomplete<SelectData<T>, false>
+        autoComplete={autoComplete}
+        data-cy={outerWrapperDataCy}
+        disableCloseOnSelect={multiple}
+        disabled={disabled}
+        getOptionLabel={getOptionLabel}
+        groupBy={groupBy}
+        isOptionEqualToValue={isOptionSelected}
+        ListboxComponent={listboxComponent}
+        ListboxProps={listboxProps}
+        loading={loading}
+        onChange={onChangeSingle}
+        onClose={onClose}
+        onInputChange={onInputChange}
+        options={options}
+        PopperComponent={renderPopper}
+        renderGroup={renderGroup}
+        renderInput={renderInput}
+        renderOption={renderOption}
+        value={singleValue}
+      />
+    );
+  }
 
-  const validatedValue = useMemo(() => validateValue(value), [validateValue, value]);
+  const onChangeMultiple = (event: SyntheticEvent, value: SelectMultipleData<T>) => {
+    suppressEvent(event);
+    externalOnChange && externalOnChange(value);
+  };
+
+  let multipleValue: SelectMultipleData<T> = [];
+  if (Array.isArray(value) && value.length > 0 && value.every(isOptionSelectable)) {
+    multipleValue = value;
+  } else {
+    logWarn("Select", `Provided value is not valid '${value}'`);
+  }
 
   return (
-    <MUIAutocomplete<T, boolean>
+    <MUIAutocomplete<SelectData<T>, true>
       autoComplete={autoComplete}
       data-cy={outerWrapperDataCy}
       disableCloseOnSelect={multiple}
@@ -253,10 +312,11 @@ const Select = <T extends any>({
       getOptionLabel={getOptionLabel}
       groupBy={groupBy}
       isOptionEqualToValue={isOptionSelected}
+      ListboxComponent={listboxComponent}
       ListboxProps={listboxProps}
       loading={loading}
-      multiple={multiple}
-      onChange={onChange}
+      multiple
+      onChange={onChangeMultiple}
       onClose={onClose}
       onInputChange={onInputChange}
       options={options}
@@ -264,14 +324,21 @@ const Select = <T extends any>({
       renderGroup={renderGroup}
       renderInput={renderInput}
       renderOption={renderOption}
-      value={validatedValue}
+      value={multipleValue}
     />
   );
 };
 
-export const SelectWithProps = Select;
-
-export default localized(Select as any, {
+export const LocalizedSelect = localized(Select, {
   dataCyShortcut: DATA_CY_SHORTCUT,
   localizableProps: LOCALIZABLE_PROPS,
-}) as <T extends any>(props: ISelect<T>) => JSX.Element;
+}) as <T extends SelectDataAllowed>(props: SelectProps<T>) => JSX.Element;
+
+/**
+ * @deprecated Select offers built-in prop virtualized
+ */
+export const SelectVirtualized = <T extends SelectDataAllowed>(props: SelectProps<T>) => (
+  <LocalizedSelect {...props} virtualized />
+);
+
+export default LocalizedSelect;