+++ src/components/Tabs/index.tsx
@@ -1,4 +1,5 @@
 import React, { FC } from "react";
+import { useTheme } from "@mui/material";
 import Box from "@mui/material/Box";
 import MUITab from "@mui/material/Tab";
 import MUITabs from "@mui/material/Tabs";
@@ -9,16 +10,17 @@ export const DATA_CY_DEFAULT = "tabs";
 
 const Tabs: FC<TabsType> = ({
   dataCy = DATA_CY_DEFAULT,
-  labelList,
   color = "primary",
   orientation = "horizontal",
   style,
+  tabList,
 }) => {
   const [value, setValue] = React.useState(0);
 
   const handleChange = (event: React.SyntheticEvent, newValue: number) => {
     setValue(newValue);
   };
+  const theme = useTheme();
   return (
     <Box sx={orientation === "vertical" ? { display: "flex" } : { display: undefined }}>
       <MUITabs
@@ -26,14 +28,28 @@ const Tabs: FC<TabsType> = ({
         style={style}
         value={value}
         onChange={handleChange}
-        textColor={color}
-        indicatorColor={color}
+        textColor="primary"
+        TabIndicatorProps={{ style: { background: theme.palette[color].main } }}
         orientation={orientation}
         variant={orientation === "vertical" ? "standard" : "scrollable"}
         scrollButtons="auto"
       >
-        {labelList.map((i, index) => {
-          return <MUITab label={i.label} wrapped={i.wrapped} key={index} />;
+        {tabList.map((i, index) => {
+          return (
+            <MUITab
+              label={i.label}
+              wrapped={i.wrapped}
+              key={index}
+              disabled={i.disabled}
+              style={{
+                color: i.textColor
+                  ? theme.palette[i.textColor].main
+                  : value === index
+                    ? theme.palette[color].main
+                    : undefined,
+              }}
+            />
+          );
         })}
       </MUITabs>
 
@@ -43,7 +59,7 @@ const Tabs: FC<TabsType> = ({
         aria-labelledby={`${orientation}-tab-${value}`}
         style={{ width: "100%" }}
       >
-        {labelList[value].content && labelList[value].content}
+        {tabList[value].content && tabList[value].content}
       </div>
     </Box>
   );